third_party/blink/web_tests/external/wpt/storage-access-api/storage-access-permission.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Permission denied state is hidden
  assert_equals: expected "prompt" but got "denied"
Harness: the test ran to completion.
